migrate preferences to local storage
