+++ src/manifest.json
@@ -52,7 +52,8 @@
     "page": "background.html"
   },
   "permissions": [
-    "messagesRead"
+    "messagesRead",
+    "storage"
   ],
   "options_ui": {
     "page": "options/options.html"
