+++ src/options/options.html
@@ -63,7 +63,7 @@
   </div>
 
   <script src="/scripts/i18n.js"></script>
-  <script src="/options/preferences.js"></script>
+  <script type="module" src="/options/preferences.js"></script>
 </body>
 
 </html>
